services/orders/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg[binary]==3.2.6
redis==5.2.1
celery==5.4.0
gunicorn==23.0.0
requests==2.32.3

